requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.1
SQLAlchemy==1.3.22
sqlalchemy-filters==0.12.0
PyMySQL==1.0.2
PyYAML==6.0
pytz==2021.3
bcrypt==3.2.2
